requirements_dvi.txt
Pillow>=9.5.0
PyYAML>=6.0
allure-pytest>=2.13.1
allure-python-commons>=2.13.1
allure-python-commons-test>=2.13.1
attrs>=22.2.0
certifi>=2022.12.7
charset-normalizer>=3.1.0
click>=8.1.3
colorama>=0.4.6
defusedxml>=0.7.1
exceptiongroup>=1.1.1
execnet>=1.9.0
idna>=3.4
iniconfig>=2.0.0
jsonpath>=0.82
lxml>=4.9.2
numpy>=1.24.2
packaging>=23.0
pandas>=2.0.0
pip>=23.0.1
pluggy>=1.0.0
py>=1.11.0
pyhamcrest>=2.0.4
pyjsonpath>=1.2.3
pytest>=7.2.2
pytest-api>=0.1.4
pytest-check>=2.1.4
pytest-html>=3.2.0
pytest-metadata>=2.0.4
pytest-xdist>=3.2.1
python-dateutil>=2.8.2
pytz>=2023.3
pyxmlparser>=0.1.6.1
requests>=2.28.2
setuptools>=67.6.1
shortuuid>=1.0.11
six>=1.16.0
tomli>=2.0.1
tzdata>=2023.3
urllib3>=1.26.15
wheel>=0.40.0
xdist>=0.0.2